+++ docs/changelog.rst
@@ -19,6 +19,8 @@ Chula v0.13.0 (dev)
 
 - Add support for using the Eventlet wsgi provider.
 - Add max-request and access-log support to :command:`chula-run`.
+- Switch to httplib2 when running bat tests.
+- Fix various http GET/POST bugs.
 
 :Source: http://github.com/jmcfarlane/chula
 
@@ -30,7 +32,7 @@ Chula v0.12.0 (latest)
 - Add support for using the Gunicorn wsgi provider.
 - Add support for using the Tornado wsgi provider.
 - Add support for "restfull" urls.  This feature is only available
-  when using the regex mapper, see: :ref:`configuration`
+  when using the regex mapper, see: :ref:`configuration`.
 - Don't include a stacktrace for 404 requests served by
   :meth:`chula.www.controller.error.Error._crappy_static_server`.